Move lib/mdx_autolink.py to common/utils/markdown.py
Also remove the target="_blank", because safe_markdown strips it out anyway, and also fuck opening links in new windows
